# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fireboy & Watergirl Multiplayer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
  completed: 'completed',
} as const;

export interface Room {
  /** 4-letter room code */
  code: string;
  /** Number of players in the room (0-2) */
  playerCount: number;
  status: RoomStatus;
  /** Current level number */
  level: number;
  createdAt: string;
}

export interface RoomInput {
  /**
     * Level to play (1-3)
     * @minimum 1
     * @maximum 3
     */
  level: number;
}

export interface ErrorResponse {
  error: string;
}