# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fireboy & Watergirl Multiplayer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Game room operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new game room
      description: Creates a new multiplayer room and returns the room code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{code}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room info
      description: Returns room status and player count
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Room:
      type: object
      properties:
        code:
          type: string
          description: 4-letter room code
        playerCount:
          type: integer
          description: Number of players in the room (0-2)
        status:
          type: string
          enum: [waiting, playing, completed]
        level:
          type: integer
          description: Current level number
        createdAt:
          type: string
          format: date-time
      required:
        - code
        - playerCount
        - status
        - level
        - createdAt

    RoomInput:
      type: object
      properties:
        level:
          type: integer
          description: Level to play (1-3)
          minimum: 1
          maximum: 3
      required:
        - level

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error