# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fireboy & Watergirl Multiplayer API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  /** 4-letter room code */
  code: string;
  /** Number of players in the room (0-2) */
  playerCount: number;
  status: RoomStatus;
  /** Current level number */
  level: number;
  createdAt: Date;
}