# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fireboy & Watergirl Multiplayer API
 * OpenAPI spec version: 0.1.0
 */

export interface RoomInput {
  /**
     * Level to play (1-3)
     * @minimum 1
     * @maximum 3
     */
  level: number;
}